requirements.txt
python-dotenv
openpyxl
pandas
webdriver-manager
selenium